ccli: Add CODING_STYLE and CONTRIBUTE documents
Now that people are starting to contribute to the project, there should be
documentation that explains how that should be done. One can not expect
people to "just know" if there's nothing telling them.

Add a CONTRIBUTE document that explains the expected way to contribute, as
well as a CODING_STYLE document that explains how the code should be.

Both were copied from trace-cmd and modified for libccli.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>