ccli: Allow for NULL words in completion list
The list returned by the completion could contain NULL pointers (failed to
allocate). The completion could either fail and clean up, or could just
ignore the failed allocation and act as if the matched word did not exist.
Sure that could be an issue for the user, but if it is hitting allocation
errors, there's probably other things the user needs to be worried about.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>