ccli: Add a new macro ISSPACE() and replace isspace() calls with it
This fixes a potential undefined behavior of ctype functions usage.
The undefined behavior happens if it receives an argument with type
signed char and the value is negative, but not EOF.

"""
int isspace(int c);

The c argument is an int, the value of which the application shall
ensure is a character representable as an unsigned char or equal to
the value of the macro EOF. If the argument has any other value, the
behavior is undefined. [1]
"""

EOF = -1
unsigned char range = [0, 255]

This means, if the value is not in range [-1, 255], ctype functions
result in undefined behavior.

Note: It's safe for the Linux kernel not to cast the argument to
unsigned char, because it internally already does it. [2]

[1]: https://pubs.opengroup.org/onlinepubs/009604499/functions/isspace.html
[2]: https://github.com/torvalds/linux/blob/v5.16/include/linux/ctype.h#L21-L39

This adds a new macro ISSPACE() that internally casts the argument to unsigned
char. Also, replace all isspace() calls with this macro.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>