ccli: Add unit tests to test the library
To prevent regressions, add unit tests to test the library.

This was inspired by the work from Tzvetomir Stoyanov who added the
CUTEST suite to libtracefs.

The unit tests utilize a pthread that does the ccli_loop() and the tests
will sends commands to it via file descriptors created by pipes. This way
the interactions between the user and the loop can be simulated.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>