ccli: Add callbacks for no command and unknown command
To give a bit more power to the developer, allow for the default
operations for both no command (the user just hitting enter) or an unknown
command (a command not registered) to have their default operations
overridden. This way the application could do something special on just an
"enter" or for bad commands (like give hints to what should have been
entered).

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>