ccli / read-file: Add ccli_history() API
Add ccli_history() to retrieve the previous commands that were done and
let the callbacks have access to it.

By doing this, read-file can now see if the previous command was a "read"
or a "dump" and if so, allow the user to just hit "enter" (by using the
ccli_register_default() function) and then the user will see the next part
of the file after what was read previously.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>