ccli: Split documentation file and add man page for ccli_history()
As the libccli-use.txt is at the max of how many functions a man page
should describe, split out the functions that are helpers for callbacks.
Namely, ccli_printf(), ccli_parse_line(), ccli_argv_free() and
ccli_history(). Note that ccli_history() was missing a description in the
man page to begin with.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>