ccli: Add ccli_execute() command
Add the API ccli_execute() that will allow a user to execute a command
outside the ccli_loop(). This could be useful if an application wants to
execute some initial commands before handing off the control to a user.

For example, perhaps saving a previous session's history and then
re-executing it on a new session.

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>