ccli: Keep current line when moving around in history
Be more like a shell and if the user types something on the command line,
and then goes through the history, do not delete what was typed, but keep
it as the last item to come back to from history.

Also, if any of the history lines are modified, the modifications will be
saved the history, similar to what other shells do.

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>