ccli: Add way to get single character input from ccli in descriptor
Add the API ccli_getchar() that reads the ccli->in. This can be useful for
commands that need to get input from the user.

Note, it only receives a single character, and it will only return
printable characters or white space. It will no return escaped characters,
but will return 0 for Ctrl^C and -1 for EOL or error.

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>