ccli: Have a way to put something other than space after matched completion
While using the library, I needed a way to have the completion logic print
something other than a space after the matched word. Since the completion
is ABI with version 1.1.0, it can not be changed. Although we can change
the const char *match to char *match without breaking binary
compatibility. Have the '\0' be a command to print something other than a
space after the matched completion.

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>