ccli: Fix HOME, END, PAGE UP and DOWN
Seems a break was removed when the read_char code was rearranged. This
caused the "num" to be tested in the same iteration of the loop as it is
set (when it should be tested in the next iteration). This broke a few
things.

Fixup the LEFT_WORD and RIGHT_WORD as that was created when things were
broken.

Fixes: c8f7653f7e9f ("ccli: Make a helper function to read command line char in loop")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>