ccli: Move completion logic out of ccli.c
Move the completion code out of ccli.c and into its own complete.c file.
More for completions can be added to it.

The refresh() function had to be changed to refresh_line() because it now
needs to be global within the library, and the history.c files has its own
refresh() function.

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>