ccli: Fix ccli_page() with line > 1
Right now the ccli_page() function expects to be called with a line of 1
when starting, where it will initialize the rows of the terminal. If a
user passes in something greater than 1, the rows will not be initialized
and a modulus will be done against rows which will be zero. This will
cause a fault in the program.

When line is greater than 1, check if rows is initialized, and if not,
initialize it then.

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>