ccli: Have ccli_file_completion only show files and not full paths
Instead of showing the full paths for files, only show the files and
directories within the current directory on the command line. This is more in
line with bash file tab completion.

That is, instead of doing a tab completion on /bin and getting a list of

  /bin/a2x                            /bin/makedtx
  /bin/a5toa4                         /bin/makeglossaries
  /bin/aa-enabled                     /bin/makeglossaries-lite
  /bin/aa-exec                        /bin/makeindex
  /bin/aa-features-abi                /bin/makejvf
  /bin/aclocal                        /bin/man
  /bin/aclocal-1.16                   /bin/man-recode
  /bin/add-patch                      /bin/mandb
  /bin/addpart                        /bin/manpage-alert

Have it just do:
  a2x                                 makedtx
  a5toa4                              makeglossaries
  aa-enabled                          makeglossaries-lite
  aa-exec                             makeindex
  aa-features-abi                     makejvf
  aclocal                             man
  aclocal-1.16                        man-recode
  add-patch                           mandb
  addpart                             manpage-alert

To accomplish this, the ccli descriptor is given a "display_index" that is used
to skip over the directory names of matched values. The ccli_file_completion()
function will set this field in the ccli descriptor, and then on printing the
list the string skips over the amount of characters specified by this field.
Then the display_index field is reset back to zero.

Currently this is only available for the ccli_file_completion() completion, but
may be later exposed to be part of the ABI.

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>