ccli completion: Simplify the command match code
An array of command names is used to find a match, and then if there's only one
found, it uses the command->cmd to insert the string. There's no reason to use
the command string directly, when there's already a commands array of strings
that found the match.

This removes the ordering dependency of the "last" variable being tied to the
command list. Also, this can be extended to merge the code path with the other
matches done after the command.

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>