ccli: Add API ccli_register_completion_table()
The normal ccli_register_completion() routines can be difficult to deal with.
Add a way to just create a table, and walk it.

When it comes to completions that are more static (like subcommands to another
command) it's easier to just statically create them, then to have a function
generate the lists.

Add a new structure:

  struct ccli_completion_table {
	const char				*name;
	ccli_completion				completion;
	const struct ccli_completion_table	*options[];
  };

The name is the name of the current word to match. The completion function is
optional and is called on the next word after the name is already matched.

To create the hierarchy, the options[] are for the next word to match after the
name has been matched. For example, if we had two commands "open" and "close"
and the "open" command had the next words be "file" and "directory" whereas
"close" had the next words be "now" or "later", we could have something like:

const struct ccli_completion_table complete_open_file = {
	.name "file",
	.completion = open_file_completion,
	.options = { NULL },
};

const struct ccli_completion_table complete_open_dir = {
	.name "dir",
	.completion = open_dir_completion,
	.options = { NULL },
};

const struct ccli_completion_table complete_open = {
	.name = "open",
	.options = { &complete_open_file, &complete_open_dir, NULL },
};

const struct ccli_completion_table complete_close_now = {
	.name "now",
	.completion = close_now_completion,
	.options = { NULL },
};

const struct ccli_completion_table complete_close_later = {
	.name "later",
	.completion = close_later_completion,
	.options = { NULL },
};

const struct ccli_completion_table complete_close = {
	.name = "close",
	.options = { &complete_close_now, &complete_close_later , NULL },
};

/* The root table isn't used, it just holds the other commands */
const struct ccli_completion_table complete_table = {
	.name = "main", // this is ignored
	.options = { &complete_open, &complete_close, NULL},
};

	ccli_register_completion_table(ccli, &complete_table, NULL);

And this is the same as calling ccli_register_completion() for the "open" and
"close" commands, but then they would need to parse for their sub commands.

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>