ccli: Add new ccli_register_command_table() API
When commands have sub commands, it makes more sense to define them statically
in a table then to have to parse them dynamically in the main command. To deal
with this, introduce a ccli_command_table structure, that will handle sub
commands (and in the future options).

The structure is:

  struct ccli_command_table {
	const char				*name;
	ccli_command_callback			command;
	const struct ccli_option_table		*options;
	const struct ccli_command_table		*subcommands[];
 };

Just like the options field in struct ccli_completion_table, the subcommands
field that lists the sub commands must be NULL terminated, and exist even if
there are no sub commands.

Note, if a command is registered with ccli_register_command() then that takes
precedence.

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>