ccli: Add default usage for command table commands
If a command has subcommands but no command callback (because it only
allows the suboptions), then there's no reason for the application to
supply one. Instead create a default "usage" message (if the application
doesn't like it, then it can create a command callback). The usage message
will show up to the failed command and the available sub commands. If an
unknown subcommand was added to by the user, it will state "unknown
option".

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>