ccli: Add ccli_register_command_delimiter() API
Add a new API ccli_register_command_delimiter() that will register a
delimiter which will be used to stop processing a line besides a newline.
After a newline is entered, if the line has this delimiter, it will process
the text up before the delimiter as a command and then the rest of the line
until it hits another delimiter or the end of the line.

This is similar to how the ";" is a delimiter to shell scripts.

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>