ccli: Add ccli_console_acquire/release() API
Add a way to put back the terminal settings that ccli setup.

This is useful for commands that fork children, and need the console put
back to what it was for the child. Before doing an exec*() on a child, the
ccli_console_release() is called to allow the child to have the same
terminal of that before the ccli application was called.

On exit of the child, ccli_console_acquire() is called to let the ccli
acquire the console again.

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>