ccli: Add line_update() to only update part of the line
Currently every time a character is typed on the command line the entire
line is rewritten. Instead, keep track of what was written and where an
update needs to be done and only write what is needed.

line_update() does this, but line_refresh() is still around to handle the
more complex cases where it is easier to just rewrite the line.

Two helper functions were added as that are used by both of those functions.
That is, go_back() which moves the cursor back and add_pad() to add padding
to the screen.

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>