ccli: Have libccli.a build off of libccli.so
Building libccli.a directly from the object files makes the __hidden
functions visible to applications that use the static library and will
conflict with functions that those applications may use if they use the same
names.

By building the libccli.a from libccli.so, it makes the hidden function weak
and should not cause issues when there's other functions being used.

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>