ccli: Free history properly
The history is freed in ccli_free() by iterating over the ccli->history[]
array up until ccli->history_size. But history_size is the number of
histories added. It may be larger than the history[] array. That's because
when the array hits ccli->history_max, it starts to loop over the array and
not add more entries.

Add a history_free() helper function in the history.c code and have the
ccli.c call that and not worry about how to actually free it.

Fixes: 406290d2 ("ccli: Free history in ccli_free()")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>