ccli: Merge completions into commands
The CCLI_DEFINE_COMMAND() and CCLI_DEFINE_COMPLETION() are basically the
same, and most of the time if one is used, then the other is too.

Consolidate them to make one command table. The CCLI_DEFINE_COMMAND() usage
is the same, but the CCLI_DEFINE_COMPLETION() now takes a "command"
parameter and creates a ccli_command_table structure.

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>