ccli: Relicense the library from LGPL-2.1 to MIT
Due to many requirements of LGPL code where every release needs to be
archived in case of someone asking for that version of the code, some
companies do not allow the code without a lot of rationale to have it.

To simplify the use of libccli, relicense the code from LGPL-2.1 to MIT.

All the files have been relicensed with the exception of the read-file.c
example code. That has a modification from an author that has not yet given
their permission to change the license.

There was modifications to the README file, but that file had no license
attached to it.

Acked-by: Steve Jothen <sjothen@gmail.com>
Acked-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Acked-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>